% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmmm.R
\name{bayesmodel}
\alias{bayesmodel}
\title{bayesmodel}
\usage{
bayesmodel(data, y, variables, observations = NULL, cores = 4,
  chains = 4, adstock_range = c(0.1, 0.7), chainlength = 2000)
}
\arguments{
\item{data}{A data frame containing a column of observations (labelled "date"), an optional column of pool names (labelled "poolname"), the dependent variable, and indepedendent variables.}

\item{y}{A string identifying the dependent variable.}

\item{variables}{a data frame (optionally created by the \code{\link{variable_grid}} function) containing one row for each independent variable and the following columns: ' \describe{
  \item{variable}{The variable name}
  \item{expected sign}{Either "+" or "-" for half-normal prior distributions. Any other values will be ignored and treated as NA. Default NA}
  \item{transformation}{Choose between "standard", "standard pooled", or "media".}
  \item{reference.point}{Choose between "0", "min", or "max".}
  \item{global}{An optional boolean to specify whether coefficients should vary across pools. If FALSE, a separate coefficient will be estimated for each pool using hierarchical partial-pooling, otherwise a single coefficient will be estimated. Default FALSE.}
  \item{prior.mean}{The prior mean. NA assumes a weakly informative prior distribution (i.e. zero after standardisation).}
  \item{prior.sd}{The prior standard deviation. NA assumes a weakly informative prior distribution (i.e. 1 after standardisation).}
}}

\item{observations}{an optional string vector of observation names, e.g. c("2014-01-01", "2014-01-08", ...). Default NULL}

\item{cores}{The number of cores to be used for the markov chain estimation. Default 4.}
}
\value{
\describe{
  \item{model}{the Stan model}
  \item{summary}{a data frame of untransformed MAP estimates}
  \item{script}{the generated Stan code}
  \item{scale}{the scaling factors used for standardisation}
  \item{y}{the y vector}
  \item{x}{the x data frame}
  \item{priors}{the priors (if any)}
  \item{observations}{the observation names (if any)}
}
}
\description{
Creates a bayesian model using rstan.
}
\examples{
model <- bayesmodel(dep, indeps, priors, observations=obs)
}
